add errors in api

+++ projectX-API/app/Http/Controllers/TutoController.php
@@ -7,6 +7,7 @@ use App\User;
 use App\Tuto;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
 
 
 class TutoController extends Controller
@@ -18,7 +19,23 @@ class TutoController extends Controller
      * @return \Illuminate\Http\Response
      */
     function store(Request $request){
-        Tuto::create([
+        $validator = Validator::make($request->all(),[
+            'title' => 'string|required | unique:tutos|max:255',
+            'difficulty' => 'required',
+            'langage' => 'required',
+            'state' => 'integer|required',
+            'author_id' => 'integer|required',
+            'summary' => 'string|required',
+            'content' => 'string|required',
+            'pathImg' => 'string|required',
+        ]);
+        if ($validator->fails()){
+            return response()->json([
+                "code" => 422,
+                'message'   =>  'Error validator']
+                ,422);
+        }
+        $newEntry = Tuto::create([
             'title'=>$request['title'],
             'difficulty'=>$request['difficulty'],
             'langage'=>$request['langage'],
@@ -32,7 +49,8 @@ class TutoController extends Controller
             'nb_ratings'=>$request['nb_ratings'],
             'files'=>$request['files'],
         ]);
-        return response()->json('Tuto create');
+
+        return response()->json('new_entry = '.$newEntry->id,201);
     }
 
     function index(Request $request){
@@ -47,6 +65,18 @@ class TutoController extends Controller
      * @return \Illuminate\Http\Response
      */
     function show($id){
+        if(intval($id) == NULL){
+            return response()->json(array(
+                'code'      =>  400,
+                'message'   =>  "Invalid query"
+            ), 400);
+        }
+        elseif(Tuto::whereId($id)->first() == null){
+            return response()->json(array(
+                'code'      =>  404,
+                'message'   =>  "Ressource not found"
+            ), 404);
+        }
         $tutos =  DB::table('tutos')->whereId($id)->first();
         return response()->json($tutos);
     }
@@ -59,6 +89,34 @@ class TutoController extends Controller
      * @return \Illuminate\Http\Response
      */
     function update(Request $request, $id){
+        if(intval($id) == NULL){
+            return response()->json(array(
+                'code'      =>  400,
+                'message'   =>  "Invalid query"
+            ), 400);
+        }
+        elseif(Tuto::whereId($id)->first() == null){
+            return response()->json(array(
+                'code'      =>  404,
+                'message'   =>  "Ressource not found"
+            ), 404);
+        }
+        $validator = Validator::make($request->all(),[
+            'title' => 'string| unique:tutos|max:255',
+            'difficulty' => '',
+            'langage' => 'string',
+            'state' => 'integer',
+
+            'summary' => 'string',
+            'content' => 'string',
+            'pathImg' => 'string',
+        ]);
+        if ($validator->fails()){
+            return response()->json([
+                'code'      =>  422,
+                'message'   =>  'Error validator']
+                ,422);
+        }
         $tuto = Tuto::whereId($id)->first();
         $tuto->update([
             'title'=>$request['title'],
@@ -68,7 +126,7 @@ class TutoController extends Controller
             'summary'=>$request['summary'],
             'content'=>$request['content'],
             'pathImg'=>$request['pathImg'],
-            
+
             'files'=>$request['files'],
         ]);
         return response()->json('Update tuto');
@@ -84,8 +142,14 @@ class TutoController extends Controller
      */
     function destroy( $id){
         $tuto = Tuto::whereId($id)->first();
+        if ($tuto == null) {
+            return response()->json(array(
+                'code'      =>  404,
+                'message'   =>  "Ressource not found"
+            ), 404);
+        }
         $tuto->delete();
-        return response()->json('Tuto deleted');
+        return response()->json('Tuto nb='.$id.' deleted');
 
     }
 }
